Refactor login and signup to be components within AuthScreen
