+++ src/components/auth/index.ts
@@ -1,0 +1,7 @@
+import LoginForm from './LoginForm';
+import SignUpForm from './SignupForm';
+
+export {
+	LoginForm,
+	SignUpForm,
+};
