+++ src/components/index.ts
@@ -6,10 +6,13 @@ import ShowCard from './ShowCard';
 import ShowCarousel from './ShowCarousel';
 import SearchInput from './SearchInput';
 import SideNavigator from './SideNavigator';
+import { LoginForm, SignUpForm } from './auth';
 
 export {
 	ShowCard,
 	ShowCarousel,
 	SearchInput,
 	SideNavigator,
+	LoginForm,
+	SignUpForm,
 };
